+++ src/duke/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class MarkCommand extends Command {
 
     private final int index;
