+++ src/duke/command/UnrecognizedCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class UnrecognizedCommand extends Command {
 
     public UnrecognizedCommand() {
