+++ src/duke/task/Deadline.java
@@ -1,8 +1,6 @@
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
+package duke.task;
+
+import duke.DukeDateTime;
 
 public class Deadline extends Task {
